+++ src/assets/icons/Right/index.tsx
@@ -1,15 +1,13 @@
-import { FC, SVGProps } from 'react';
+import { SVGProps } from 'react';
 
-const RightIcon: FC<SVGProps<SVGSVGElement>> = (
-  props: SVGProps<SVGSVGElement>,
-): JSX.Element => (
-  <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 20 20" {...props}>
-    <path
-      fillRule="evenodd"
-      d="M7.21 14.77a.75.75 0 01.02-1.06L11.168 10 7.23 6.29a.75.75 0 111.04-1.08l4.5 4.25a.75.75 0 010 1.08l-4.5 4.25a.75.75 0 01-1.06-.02z"
-      clipRule="evenodd"
-    />
-  </svg>
-);
-
-export default RightIcon;
+export default function RightIcon(props: SVGProps<SVGSVGElement>) {
+  return (
+    <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 20 20" {...props}>
+      <path
+        fillRule="evenodd"
+        d="M7.21 14.77a.75.75 0 01.02-1.06L11.168 10 7.23 6.29a.75.75 0 111.04-1.08l4.5 4.25a.75.75 0 010 1.08l-4.5 4.25a.75.75 0 01-1.06-.02z"
+        clipRule="evenodd"
+      />
+    </svg>
+  );
+}