+++ src/lib/api/finnhub/Finnhub.ts
@@ -3,6 +3,6 @@ import axios from 'axios';
 export const Finnhub = axios.create({
   baseURL: 'https://finnhub.io/api/v1',
   headers: {
-    'X-Finnhub-Token': process.env.FINNHUB_KEY,
+    'X-Finnhub-Token': import.meta.env.VITE_FINNHUB_KEY,
   },
 });