+++ src/lib/api/finnhub/index.ts
@@ -1,5 +1,0 @@
-export const finnhub = {
-  base: await import('./Finnhub'),
-  getQuotes: (await import('./quote')).quote,
-  getTrends: (await import('./trends')).getTrends,
-};