+++ src/lib/api/finnhub/trends.ts
@@ -1,4 +1,4 @@
-import Finnhub from './Finnhub';
+import { Finnhub } from './Finnhub';
 
 export const getTrends = async (symbol: string) =>
   await Finnhub.get(`/stock/recommendation?symbol=${symbol}`).then(