+++ src/main.tsx
@@ -1,11 +1,11 @@
 import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
-import './index.css';
 
-import Layout from 'route/layout';
+import App from './app';
+import './index.css';
 
 createRoot(document.getElementById('root') as HTMLElement).render(
   <StrictMode>
-    <Layout />
+    <App />
   </StrictMode>,
 );