+++ src/route/app/dividends/layout.tsx
@@ -1,16 +1,13 @@
-import { FC } from 'react';
-import Calendar from './(container)/calendar/container';
-import Form from './(container)/form/container';
-import List from './(container)/list/container';
+import CalendarContainer from './(container)/calendar/container';
+import FormContainer from './(container)/form/container';
+import ListContainer from './(container)/list/container';
 
-const Layout: FC = (): JSX.Element => {
+export default function Layout() {
   return (
     <div className="w-full h-full grid grid-cols-6 grid-rows-6 bg-slate-700 gap-4 p-2 lining-nums tabular-nums text-slate-200">
-      <List />
-      <Calendar />
-      <Form />
+      <ListContainer />
+      <CalendarContainer />
+      <FormContainer />
     </div>
   );
-};
-
-export default Layout;
+}