+++ src/route/app/stocks/layout.tsx
@@ -1,7 +1,5 @@
-import { FC } from 'react';
-
-const Layout: FC = (): JSX.Element => (
-  <div className="w-full h-full grid grid-cols-2 grid-rows-1 bg-slate-700 gap-4 p-2 lining-nums tabular-nums text-slate-200"></div>
-);
-
-export default Layout;
+export default function Layout() {
+  return (
+    <div className="w-full h-full grid grid-cols-2 grid-rows-1 bg-slate-700 gap-4 p-2 lining-nums tabular-nums text-slate-200"></div>
+  );
+}