+++ src/utils/DayJs.ts
@@ -1,11 +1,10 @@
-import dayjs, { Dayjs } from 'lib/dayjs';
-import 'lib/dayjs/locale/en';
-import businessDays from 'lib/dayjs/plugin/businessDays';
+import dayjs, { Dayjs } from '@Lib/dayjs';
+import businessDays from '@Lib/dayjs/plugin/businessDays';
+import '@Lib/dayjs/locale/en';
 
 dayjs.extend(businessDays);
 dayjs.locale('en');
 
-const DayJs = (date?: String): Dayjs =>
-  date ? dayjs(date.split('/').reverse().join('-')) : dayjs();
-
-export default DayJs;
+export default function DayJs(date?: String): Dayjs {
+  return date ? dayjs(date.split('/').reverse().join('-')) : dayjs();
+}